vkr/backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.0,<5.4.0
psycopg2-binary>=2.9.9,<3.0.0
whitenoise>=6.6.0,<6.7.0
django-cors-headers>=4.3.1,<4.4.0
gunicorn>=21.2.0,<21.3.0
celery>=5.3.6,<5.4.0
redis>=5.0.1,<5.1.0
drf-yasg
django-filter>=23.5,<23.6
Pillow
requests